graphworker/requirements.txt
networkx
psycopg2-binary
fastapi
uvicorn
python-dotenv
